make sure the right window has the focus when typing in the password
When a choice dialog is shown, the active window may be changed after
closing the dialog (i.e. if the user first does something else, then
selects a password from the dialog). Abort the script, if the window id
as xdotool reports it has changed since reading the window title.